+++ ui/src/LoginPage.tsx
@@ -5,9 +5,10 @@ import { useAuth } from "./AuthContext";
 
 interface LoginProps {
     auth: boolean;
+    error?: undefined | string;
 }
 
-const LoginPage: React.FC<LoginProps> = ({ auth }) => {
+const LoginPage: React.FC<LoginProps> = ({ auth, error }) => {
     const [isLoading, setIsLoading] = useState(false);
     const { login } = useAuth();
     const navigate = useNavigate();
@@ -36,7 +37,7 @@ const LoginPage: React.FC<LoginProps> = ({ auth }) => {
                 <div className="text-center text-7xl font-bold text-white mb-20 z-10">Cosmian KMS user interface</div>
                 <div className="space-y-6 text-center w-1/2">
                     {auth && <p className="text-white">Sign up for free and explore rights delegation for multiple users</p>}
-
+                    {error && <p className="text-purple-700">{error}</p>}
                     {isLoading ? (
                         <Spin size="large" />
                     ) : auth ? (
@@ -48,8 +49,6 @@ const LoginPage: React.FC<LoginProps> = ({ auth }) => {
                             ACCESS KMS
                         </Button>
                     )}
-                    {/* Accessible Error Message */}
-                    <div id="error-message" role="alert" aria-live="polite" className="text-red-500 text-sm mt-2"></div>
                 </div>
             </div>
         </div>